feat(auth) : implement a mandatory login first token based authentication

+++ controllers/expense.delete.js
@@ -5,6 +5,9 @@ import db from '../db/database.connect.js';
 export default wrapper(async (req , res) => {
     const data = await parseBody(req);
 
+    if(!data || !data.expense_id){
+        throw new Error('details are not provided')
+    }
     const deleteExpense = `DELETE FROM ${process.env.EXPENSE_TABLE_NAME} WHERE expense_id = ?`;
 
     db.query(deleteExpense , [data.expense_id]);
